requirements.txt
streamlit>=1.38
pandas>=2.2
openpyxl>=3.1
reportlab>=4.2
matplotlib>=3.8
anthropic>=0.39
python-dotenv>=1.0
